Test: Move more pure Expr/Prop related tests
